Specialises Store implementation between RocksDB and in-memory
Will be useful for optimizations
